backend/requirements.txt
fastapi
uvicorn[standard]
mcp
ollama
faiss-cpu
sentence-transformers
pdfplumber
numpy
requests
beautifulsoup4
pydantic